// nodes/Vatnode/countries.generated.ts
export interface CountryEntry {
	code: string;
	name: string;
	pattern: string | null;
	viesEligible: boolean;
}

export const COUNTRIES: CountryEntry[] = [
	{
		"code": "AD",
		"name": "Andorra",
		"pattern": "^AD[A-Z]\\d{6}[A-Z]$",
		"viesEligible": false
	},
	{
		"code": "AL",
		"name": "Albania",
		"pattern": "^AL[A-Z]\\d{8}[A-Z]$",
		"viesEligible": false
	},
	{
		"code": "AT",
		"name": "Austria",
		"pattern": "^ATU\\d{8}$",
		"viesEligible": true
	},
	{
		"code": "BA",
		"name": "Bosnia and Herzegovina",
		"pattern": "^BA\\d{12}$",
		"viesEligible": false
	},
	{
		"code": "BE",
		"name": "Belgium",
		"pattern": "^BE[01]\\d{9}$",
		"viesEligible": true
	},
	{
		"code": "BG",
		"name": "Bulgaria",
		"pattern": "^BG\\d{9,10}$",
		"viesEligible": true
	},
	{
		"code": "CH",
		"name": "Switzerland",
		"pattern": "^CHE-?\\d{3}\\.?\\d{3}\\.?\\d{3}[ ]?(MWST|^TVA|^IVA)?$",
		"viesEligible": false
	},
	{
		"code": "CY",
		"name": "Cyprus",
		"pattern": "^CY\\d{8}[A-Z]$",
		"viesEligible": true
	},
	{
		"code": "CZ",
		"name": "Czech Republic",
		"pattern": "^CZ\\d{8,10}$",
		"viesEligible": true
	},
	{
		"code": "DE",
		"name": "Germany",
		"pattern": "^DE\\d{9}$",
		"viesEligible": true
	},
	{
		"code": "DK",
		"name": "Denmark",
		"pattern": "^DK\\d{8}$",
		"viesEligible": true
	},
	{
		"code": "EE",
		"name": "Estonia",
		"pattern": "^EE\\d{9}$",
		"viesEligible": true
	},
	{
		"code": "EL",
		"name": "Greece",
		"pattern": "^EL\\d{9}$",
		"viesEligible": true
	},
	{
		"code": "ES",
		"name": "Spain",
		"pattern": "^ES[A-Z0-9]\\d{7}[A-Z0-9]$",
		"viesEligible": true
	},
	{
		"code": "FI",
		"name": "Finland",
		"pattern": "^FI\\d{8}$",
		"viesEligible": true
	},
	{
		"code": "FR",
		"name": "France",
		"pattern": "^FR[A-HJ-NP-Z0-9]{2}\\d{9}$",
		"viesEligible": true
	},
	{
		"code": "GB",
		"name": "United Kingdom",
		"pattern": "^GB(\\d{9}|^GB\\d{12}|^GD\\d{3}|^HA\\d{3})$",
		"viesEligible": false
	},
	{
		"code": "GE",
		"name": "Georgia",
		"pattern": "^GE\\d{9}$",
		"viesEligible": false
	},
	{
		"code": "GR",
		"name": "Greece",
		"pattern": "^EL\\d{9}$",
		"viesEligible": true
	},
	{
		"code": "HR",
		"name": "Croatia",
		"pattern": "^HR\\d{11}$",
		"viesEligible": true
	},
	{
		"code": "HU",
		"name": "Hungary",
		"pattern": "^HU\\d{8}$",
		"viesEligible": true
	},
	{
		"code": "IE",
		"name": "Ireland",
		"pattern": "^IE\\d{7}[A-W][A-IW]?$|^IE\\d[A-Z+*]\\d{5}[A-W]$",
		"viesEligible": true
	},
	{
		"code": "IS",
		"name": "Iceland",
		"pattern": "^IS\\d{5,6}$",
		"viesEligible": false
	},
	{
		"code": "IT",
		"name": "Italy",
		"pattern": "^IT\\d{11}$",
		"viesEligible": true
	},
	{
		"code": "LI",
		"name": "Liechtenstein",
		"pattern": "^LI\\d{5}$",
		"viesEligible": false
	},
	{
		"code": "LT",
		"name": "Lithuania",
		"pattern": "^LT(\\d{9}|^LT\\d{12})$",
		"viesEligible": true
	},
	{
		"code": "LU",
		"name": "Luxembourg",
		"pattern": "^LU\\d{8}$",
		"viesEligible": true
	},
	{
		"code": "LV",
		"name": "Latvia",
		"pattern": "^LV\\d{11}$",
		"viesEligible": true
	},
	{
		"code": "MC",
		"name": "Monaco",
		"pattern": "^FR[A-HJ-NP-Z0-9]{2}\\d{9}$",
		"viesEligible": false
	},
	{
		"code": "MD",
		"name": "Moldova",
		"pattern": "^MD\\d{7}$",
		"viesEligible": false
	},
	{
		"code": "ME",
		"name": "Montenegro",
		"pattern": "^ME\\d{8}$",
		"viesEligible": false
	},
	{
		"code": "MK",
		"name": "North Macedonia",
		"pattern": "^MK\\d{13}$",
		"viesEligible": false
	},
	{
		"code": "MT",
		"name": "Malta",
		"pattern": "^MT\\d{8}$",
		"viesEligible": true
	},
	{
		"code": "NL",
		"name": "Netherlands",
		"pattern": "^NL\\d{9}B\\d{2}$",
		"viesEligible": true
	},
	{
		"code": "NO",
		"name": "Norway",
		"pattern": "^NO\\d{9}MVA$",
		"viesEligible": false
	},
	{
		"code": "PL",
		"name": "Poland",
		"pattern": "^PL\\d{10}$",
		"viesEligible": true
	},
	{
		"code": "PT",
		"name": "Portugal",
		"pattern": "^PT\\d{9}$",
		"viesEligible": true
	},
	{
		"code": "RO",
		"name": "Romania",
		"pattern": "^RO\\d{2,10}$",
		"viesEligible": true
	},
	{
		"code": "RS",
		"name": "Serbia",
		"pattern": "^RS\\d{9}$",
		"viesEligible": false
	},
	{
		"code": "SE",
		"name": "Sweden",
		"pattern": "^SE\\d{10}01$",
		"viesEligible": true
	},
	{
		"code": "SI",
		"name": "Slovenia",
		"pattern": "^SI\\d{8}$",
		"viesEligible": true
	},
	{
		"code": "SK",
		"name": "Slovakia",
		"pattern": "^SK\\d{10}$",
		"viesEligible": true
	},
	{
		"code": "TR",
		"name": "Turkey",
		"pattern": "^TR\\d{10}$",
		"viesEligible": false
	},
	{
		"code": "UA",
		"name": "Ukraine",
		"pattern": "^UA\\d{9}$",
		"viesEligible": false
	},
	{
		"code": "XI",
		"name": "Northern Ireland",
		"pattern": "^XI(\\d{9}|^XI\\d{12}|^XI(GD|^HA)\\d{3})$",
		"viesEligible": true
	},
	{
		"code": "XK",
		"name": "Kosovo",
		"pattern": "^XK\\d{9}$",
		"viesEligible": false
	}
];

export const COUNTRY_BY_CODE: Record<string, CountryEntry> = Object.fromEntries(
	COUNTRIES.map((country) => [country.code, country]),
);